requirements.txt
pdfplumber>=0.11.0
pypdf>=4.0.0
anthropic>=0.42.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
pymupdf>=1.24.0
pytesseract>=0.3.10
supabase>=2.0.0
voyageai>=0.3.0
openai>=1.0.0
pytest>=7.0
httpx>=0.27.0
